Replace Actionable + AsMut with Action + reference counting
To get the "shared structure" we want, where applying an action only
clones as much as necessary, we'll want reference-counting. This way, if
we only update, e.g., the Table of a Game, the existing Selection will
be shared with the new Game. And if we update a Game with no clones,
we'll automatically reuse the existing game rather than cloning and
dropping.