Clean up all existing rules
This commit is several months in the making, so I don't remember my whole
thought process. The goal was to cleanly implement rules for the table actions.
There are still two actions that don't have rules, but I think I want to commit
before a possible refactor of action handling.