Replace `use super::` with `use crate::...`
When `super` refers to a module in a different file we should prefer
crate-relative (i.e., absolute) imports. This will make it clearer what's being
imported, and also ensure that modules are resilient to being moved.

Modules in the same file as their super module -- e.g., test modules -- can
continue to use `use super::` as it will be clear what they refer to.